Fixing make/with-collector so that this function should be more complient
Collector functions (as by make-collector) no longer return their arguments
which makes their item list dynamic-extent.

From the spec:
The value of a rest parameter is permitted, but not required, to share
structure with the last argument to apply.

We saw no actual bugs with this, but it seemed prudent to fix it before
we did